Enable AggregateDataExtract creation from dict
